글 목록 화면 댓글 개수 에러 수정

+++ src/main/java/com/ohoon/board/app/dto/PostListDto.java
@@ -18,13 +18,13 @@ public class PostListDto {
 
     private String title;
 
-    private int totalComments;
+    private long totalComments;
 
     private String author;
 
     private Long view;
 
-    private int totalVotes;
+    private long totalVotes;
 
     private LocalDateTime createdDate;
